% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_incubation.R
\name{compute_incubation}
\alias{compute_incubation}
\title{Compute the empirical incubation dist.
Can take into account uncertain dates of exposure.}
\usage{
compute_incubation(dates_exposure, date_onset)
}
\arguments{
\item{dates_exposure}{list containing the exposure dates. each element can be a vector of several possible exposure dates.}

\item{date_onset}{list containing the exposure dates. each element can be a vector of several dates.}
}
\value{
a data frame containing a column with the different incubation periods and a column containing their relative frequency
}
\description{
Compute the empirical incubation dist.
Can take into account uncertain dates of exposure.
}
\author{
Flavio Finger, \email{flavio.finger@lshtm.ac.uk}
}
