% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_incubation.R
\name{empirical_incubation_dist}
\alias{empirical_incubation_dist}
\title{Function to extract the empirical incubation period distribution from line list data.
Can take into account uncertain dates of exposure.}
\usage{
empirical_incubation_dist(x, dates_exposure, date_of_onset)
}
\arguments{
\item{x}{the linelist data (data.frame or linelist object) containing at least a column containing the exposure dates and one containing the onset dates. For exposure dates, each element can be a list containing several possible exposure dates.}

\item{dates_exposure}{the name of the column containing the exposure dates.}

\item{dates_exposure}{the name of the column containing the onset dates}
}
\value{
a data frame containing a column with the different incubation periods and a column containing their relative frequency
}
\description{
Function to extract the empirical incubation period distribution from line list data.
Can take into account uncertain dates of exposure.
}
\author{
Flavio Finger, \email{flavio.finger@lshtm.ac.uk}
}
