% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_names.R
\name{hash_names}
\alias{hash_names}
\title{Anonymise data using SHA1}
\usage{
hash_names(..., size = 6, full = TRUE)
}
\arguments{
\item{...}{Data fields to be hashed.}

\item{size}{The number of characters retained in the hash.}

\item{full}{A logical indicating if the a full output should be returned as a
\code{data.frame}, including original labels, shortened hash, and full
hash.}
}
\description{
This function uses SHA1 algorithm to anonymise data, based on pre-specified
data fields. Data fields are concatenated first, then each entry is
hashed. The function can either return a full detailed output, or short
anonymous labels.
}
\examples{

first_name <- c("Jane", "Joe", "Raoul")
last_name <- c("Doe", "Smith", "Dupont")
age <- c(25, 69, 36)

hash_names(first_name, last_name, age)

hash_names(first_name, last_name, age,
           size = 8, full = FALSE)

}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
